cleanup and guard against unsaved projects
