+++ Gum/Plugins/InternalPlugins/TreeView/ElementTreeViewManager.RightClick.cs
@@ -499,17 +499,37 @@ public partial class ElementTreeViewManager
 
     private void HandleImportBehavior(object sender, EventArgs args)
     {
-        _dialogService.Show<ImportBehaviorDialog>();
+        if (GuardProjectSaved("before importing behaviors"))
+        {
+            _dialogService.Show<ImportBehaviorDialog>();
+        }
     }
 
     public void ImportScreenClick(object sender, EventArgs e)
     {
-        _dialogService.Show<ImportScreenDialog>();
+        if (GuardProjectSaved("before importing screens"))
+        {
+            _dialogService.Show<ImportScreenDialog>();
+        }
     }
 
     public void ImportComponentsClick(object sender, EventArgs e)
     {
-        _dialogService.Show<ImportComponentDialog>();
+        if (GuardProjectSaved("before importing components"))
+        {
+            _dialogService.Show<ImportComponentDialog>();
+        }
+    }
+
+    private bool GuardProjectSaved(string? reason = null)
+    {
+        if (ObjectFinder.Self.GumProjectSave == null || string.IsNullOrEmpty(ProjectManager.Self.GumProjectSave.FullFileName))
+        {
+            _dialogService.ShowMessage("You must first save the project");
+            return false;
+        }
+
+        return true;
     }
 
     private void HandleAddLinkedComponentClick(object sender, EventArgs e)
